raii for attach and context next
