+++ libs/ledger/include/ledger/upow/synergetic_contract.hpp
@@ -18,6 +18,7 @@
 //------------------------------------------------------------------------------
 
 #include "chain/address.hpp"
+#include "ledger/chaincode/contract_context.hpp"
 #include "ledger/upow/synergetic_base_types.hpp"
 #include "vm/analyser.hpp"
 #include "vm/common.hpp"
@@ -54,29 +55,7 @@ struct Variant;
 
 namespace ledger {
 
-class StateAdapter;
 class StorageInterface;
-class TokenContract;
-
-struct SynergeticContractContext
-{
-  SynergeticContractContext()                                  = default;
-  SynergeticContractContext(SynergeticContractContext const &) = default;
-  SynergeticContractContext(SynergeticContractContext &&)      = default;
-  SynergeticContractContext &operator=(SynergeticContractContext const &) = default;
-  SynergeticContractContext &operator=(SynergeticContractContext &&) = default;
-
-  SynergeticContractContext(TokenContract *token_contract_param, chain::Address address,
-                            StateAdapter *state_adapter_param)
-    : token_contract{token_contract_param}
-    , contract_address{std::move(address)}
-    , state_adapter{state_adapter_param}
-  {}  //???uninline
-
-  TokenContract *token_contract{nullptr};
-  chain::Address contract_address{};
-  StateAdapter * state_adapter{nullptr};
-};
 
 class SynergeticContract
 {
@@ -106,15 +85,8 @@ public:
   void Attach(StorageInterface &storage);
   void Detach();
 
-  void updateContractContext(SynergeticContractContext context)  //???uninline
-  {
-    context_ = std::move(context);
-  }
-
-  SynergeticContractContext const &context() const  //???uninline
-  {
-    return context_;
-  }
+  void                   updateContractContext(ContractContext context);
+  ContractContext const &context() const;
 
   /// @name Actions to be taken on the synergetic contract
   /// @{
@@ -149,10 +121,10 @@ private:
   std::string objective_function_;
   std::string clear_function_;
 
-  StorageInterface *        storage_{nullptr};
-  VariantPtr                problem_;
-  VariantPtr                solution_;
-  SynergeticContractContext context_{};
+  StorageInterface *storage_{nullptr};
+  VariantPtr        problem_;
+  VariantPtr        solution_;
+  ContractContext   context_{};
 };
 
 char const *ToString(SynergeticContract::Status status);
