+++ libs/ledger/src/chaincode/contract.cpp
@@ -25,13 +25,6 @@
 namespace fetch {
 namespace ledger {
 
-ContractContext::ContractContext(TokenContract *token_contract_param, chain::Address address,
-                                 StateAdapter *state_adapter_param)
-  : token_contract{token_contract_param}
-  , contract_address{std::move(address)}
-  , state_adapter{state_adapter_param}
-{}
-
 /**
  * Dispatch the initialisation handler
  *
