+++ libs/ledger/src/executor.cpp
@@ -22,6 +22,7 @@
 #include "core/macros.hpp"
 #include "core/mutex.hpp"
 #include "ledger/chaincode/contract.hpp"
+#include "ledger/chaincode/contract_context.hpp"
 #include "ledger/chaincode/token_contract.hpp"
 #include "ledger/consensus/stake_update_interface.hpp"
 #include "ledger/executor.hpp"
